requirements.txt
fastapi
uvicorn
pdf2image
pytesseract
Pillow